nci-agency/anet#1903: Refactor the various output formats
